[--IMPORTANT--][PLEASE READ]
GUYS, please make sure the profile pictures which can be found in assets/uploads ARE in SMALL caps and that Eclipse MUST BE able to recognize the images. I pushed this change before but I don't know why apparently some of you are stuck with old ones that have CAPITALIZED names.
Please pull asap.